Remove javadocs and param names for synthetics
Due to how the SRG names are populated, the initial import of MCP data
duplicated data between the originally documented method and its synthetic
 bouncer (because type erasure and covariant return types).

Because this cannot be 'fixed' without modification to the SRG spec (which
would undoubtedly be rejected because of the nicheness), Compass now has a
sanitize staging data task. For now, this task does one job: remove javadocs
and param names from synthetic fields, methods, and their params.