Use Enigma fatjar
Enigma 1.4.2 onwards does not publish the dependency information for the
regular JAR in the POM, presumably because those would be resolved by
Gradle even if the fatjar was used.
And because for the life of me I couldn't figure out how to tell Gradle
to use the dep information available in the .module, it's easier to
use the fatjar instead.