Update WorldSelectionList.mapping
Changed the parameter supplier from the method refreshList to filterLevelSupplier as the supplied string is used to filter out levels which their names or ids do not contain the supplied string.

Also added the parameter for ARG 8 in the constructor for WorldSelectionList to match the parameter from the method refreshList.