Remove erroneous param for BlockItem#shouldPrintOpWarning
This method is cascaded from its superclass, which causes both its
parameters to be named "player" in the generated exports (the first
parameter from this mapping, the second parameter from the superclass
mapping) and thus a recompilation error.

The general fix would be to ensure that the generated export cannot
duplicate parameters in the same method, and to check for this in
validation, but the quickest fix is to remove this mapping.